model/requirements.txt
google-cloud-bigquery[bqstorage,pandas]==1.28.0
SQLAlchemy==1.3.19
mysqlclient==2.0.1
pandas==1.1.2
numpy==1.19.2
psycopg2-binary==2.8.6
web3==5.12.1
py-etherscan-api==0.8.0
tqdm==4.50.0
pymc3==3.9.3
